CXX编译选项

+++ Demo1/math/MathFunctions.cpp
@@ -1,7 +1,7 @@
 #include "MathFunctions.h"
 
 double power(double base, int exponent){
-	int result = base;
+	double result = base;
 	for (int i = 1; i < exponent; ++i)
 		result *= base;
 	return result;
